Add restriction of updating/delete map based on map owner

+++ routes/maps.js
@@ -10,6 +10,7 @@ const router  = express.Router();
 const mapQueries = require('../db/queries/maps');
 const locationQueries = require('../db/queries/locations');
 const favouriteQueries = require('../db/queries/favourites');
+const e = require('express');
 
 router.use((req,res,next) => {
   const userId = req.cookies['user_id'];
@@ -74,8 +75,15 @@ router.post('/:map_id/update', (req, res) => {
     req.body.lat,
     req.body.lng
   ];
-  mapQueries.updateMap(mapInfo)
-    .then (() => res.redirect(`/maps/${req.params.map_id}`))
+  mapQueries.getMapsById(mapInfo[0])
+  .then ((map) => {
+    if (map.owner_id == req.cookies['user_id']) {
+      mapQueries.updateMap(mapInfo)
+        .then (() => res.redirect(`/maps/${req.params.map_id}`))
+    } else {
+      res.status(401).send('Cannot update: You are not the owner of this map');
+    }
+  })
 });
 
 //Post request to delete map
@@ -83,8 +91,15 @@ router.post('/:map_id/delete', (req, res) => {
   const mapId = [
     req.params.map_id,
   ];
-  mapQueries.deleteMap(mapId)
-    .then (() => res.redirect(`/maps/`))
+  mapQueries.getMapsById(mapId[0])
+  .then ((map) => {
+    if (map.owner_id == req.cookies['user_id']) {
+      mapQueries.deleteMap(mapId)
+        .then (() => res.redirect(`/maps/`))
+    } else {
+      res.status(401).send('Cannot update: You are not the owner of this map');
+    }
+  })
 });
 
 //Put favourited maps to specific user
@@ -160,8 +175,15 @@ router.post('/:map_id/locations/:location_id', (req, res) => {
     req.body.description,
     req.body.locationImage
   ];
-  locationQueries.updateLocation(locationInfo)
-    .then (() => res.redirect(`/maps/${req.params.map_id}/locations/${req.params.location_id}`))
+  locationQueries.getLocationsById(locationInfo[0])
+    .then ((location) => {
+      if (location.creator_id == req.cookies['user_id']) {
+        locationQueries.updateLocation(locationInfo)
+        .then (() => res.redirect(`/maps/${req.params.map_id}`))
+      } else {
+        res.status(401).send('Cannot update: You are not the owner of this location');
+      }
+    })
 });
 
 //Post request to delete a location marker on a map
@@ -169,8 +191,15 @@ router.post('/:map_id/locations/:location_id/delete', (req, res) => {
   const deleteLocation = [
     req.params.location_id,
   ];
-  locationQueries.deleteLocation(deleteLocation)
-    .then (() => res.redirect(`/maps/${req.params.map_id}`))
+  locationQueries.getLocationsById(deleteLocation[0])
+  .then ((location) => {
+    if (location.creator_id == req.cookies['user_id']) {
+      locationQueries.deleteLocation(deleteLocation)
+      .then (() => res.redirect(`/maps/${req.params.map_id}`))
+    } else {
+      res.status(401).send('Cannot delete: You are not the owner of this location');
+    }
+  })
 });
 
 module.exports = router;
